+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.1.0 - Add ability to jump while tree-view is focused
 * Should work with 'shift-enter' or 'f' bindings in vim-mode!
+* Alert users about current atom release's performance issues.
 
 ## 4.0.0 - New Architecture + Jump to Tabs!
 * Add ability to jump to tabs!